Move toolset to own component
